Minor Refactoring
- Removed unused/unnecessary files, imports, functions, etc.
- Moved imports to keep a consistent ordering.
- Some other minor changes for consistency & readability of code.

+++ src/Components/File Modal Components/FileDeleteDialog.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import {
     AlertDialog,
     AlertDialogOverlay,
@@ -8,40 +6,33 @@ import {
     AlertDialogBody,
     AlertDialogFooter,
     Button,
-  } from '@chakra-ui/react'
-
+} from '@chakra-ui/react'
 
 const FileDeleteDialog = (props) => {
-    
-  const cancelRef = React.useRef()
-
     return (
         <AlertDialog
             isOpen={props.isOpen}
-            leastDestructiveRef={props.cancelRef}
             onClose={props.onClose}
         >
-        <AlertDialogOverlay>
-            <AlertDialogContent bgColor='gray.700' color='white'>
-            <AlertDialogHeader fontSize='lg' fontWeight='bold'>
-                Delete File {props.fileNo}?
-            </AlertDialogHeader>
-
-            <AlertDialogBody whiteSpace='pre-line'>
-                Are you sure you want to delete this file?{"\n"}
-                This action cannot be undone.
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-                <Button colorScheme='whiteAlpha' ref={props.cancelRef} onClick={props.onClose}>
-                Cancel
-                </Button>
-                <Button colorScheme='red' onClick={()=>{props.onClose(); props.deleteFile()}} ml={3}>
-                Delete
-                </Button>
-            </AlertDialogFooter>
-            </AlertDialogContent>
-        </AlertDialogOverlay>
+            <AlertDialogOverlay>
+                <AlertDialogContent bgColor='gray.700' color='white'>
+                    <AlertDialogHeader fontSize='lg' fontWeight='bold'>
+                        Delete File {props.fileNo}?
+                    </AlertDialogHeader>
+                    <AlertDialogBody whiteSpace='pre-line'>
+                        Are you sure you want to delete this file?{'\n'}
+                        This action cannot be undone.
+                    </AlertDialogBody>
+                    <AlertDialogFooter>
+                        <Button colorScheme='whiteAlpha' onClick={props.onClose}>
+                            Cancel
+                        </Button>
+                        <Button colorScheme='red' onClick={()=>{props.onClose(); props.deleteFile()}} ml={3}>
+                            Delete
+                        </Button>
+                    </AlertDialogFooter>
+                </AlertDialogContent>
+            </AlertDialogOverlay>
         </AlertDialog>
     )
 }
